supervisorctl avail doesn't use process/group name
`supervisorctl avail` will always print a list of all known process groups and their status, so passing the `@resource[:name]` parameter to it doesn't have any effect. Instead, we can include the parameter in the regex being used, so we can use the output for that specific process group. 

(tested using supervisor 3.3.1)